+++ changelog.md
@@ -7,7 +7,7 @@
 - `prelude` now works with the JavaScript target.
 
 - Added `ioutils` module containing `duplicate` and `duplicateTo` to duplicate `FileHandle` using C function `dup` and `dup2`.
-
+- Added `almostEqual` in `math` for comparing two float values using machine epsilon.
 
 ## Language changes